repository error handler and some method rename

+++ server/repositories/CoSupervisorThesisRepository.js
@@ -1,62 +1,45 @@
 'use strict';
 
-const sqlite = require('sqlite3');
-
-//==================================Support==================================
-
 const db = require('./db')
+
+
+//==================================Create==================================
+
 /**
  * Insert an entry into COSUPERVISORTHESIS table without performing any checks over parameters exis
  * @param {*} id_thesis must exists in THESIS 
  * @param {*} id_theacher must exists in TEACHER
  * @param {*} id_cosupervisor must exists in COSUPERVISOR
- * @returns 3-tuple of the ids
+ * @returns true
  */
 exports.addCoSupervisorThesis = (id_thesis, id_theacher, id_cosupervisor) => {
+    if(!id_thesis || id_thesis<0 || !id_theacher || id_theacher<0 || !id_cosupervisor || id_cosupervisor<0)
+        throw {error: "id must be greather than 0"}
     const sql = "INSERT INTO CoSupervisorThesis(id_thesis, id_teacher, id_cosupervisor) VALUES (?, ?, ?)"
-
     return new Promise((resolve, reject) => {
         db.run(sql, [id_thesis, id_theacher, id_cosupervisor], (err) => {
             if (err)
-                reject(err)
+                return reject({error: err.message})
             resolve(true)
         })
     })
 }
 
-/**
- * Support function to find the thesisId given the CoSupervisor id
- * @param {*} id: id of the co-supervisor
- * @returns idThesis: integer
- */
-exports.findThesisByCoSupervisorId = (id) => {
-    let idsThesis = [];
-    const sqlIdThesis = "SELECT id_thesis FROM CoSupervisorThesis WHERE id_cosupervisor = ?";
-    return new Promise((resolve, reject) => {
-        db.all(sqlIdThesis, [id], (err, rows) => {
-            if (err) {
-                reject(err);
-                return;
-            }
-            rows.map((e) => {
-                idsThesis.push(e.id_thesis);
-            });
-            resolve(idsThesis);
-        });
-    });
-}
+//==================================Get==================================
 
 /**
  * Given a selected thesis the ids of the supervisor and cosupervisors will be returned
  * @param {*} id of a selected thesis 
- * @returns [id1, id2, ...]
+ * @returns [{idTeacher: id}, {idCoSupervisor: id}, ...]
  */
-exports.findCoSupervisorIdsByThesisId = (id) => {
+exports.getIdsByThesisId = (id) => {
+    if(!id || id<0)
+        throw {error: "id must be greather than 0"}
     const sqlCoSupervisor = "SELECT id_teacher, id_cosupervisor FROM CoSupervisorThesis WHERE id_thesis = ?";
     return new Promise((resolve, reject) => {
         db.all(sqlCoSupervisor, [id], (err, rows) => {
             if (err) {
-                reject(err);
+                reject({error: err.message});
                 return;
             }
             const res = rows.map((e) => {
@@ -70,6 +53,10 @@ exports.findCoSupervisorIdsByThesisId = (id) => {
     });
 }
 
+//==================================Set==================================
+
+//==================================Delete==================================
+
 /**
  * Delete coSupervisor given the thesis id
  * @param {*} thesisId 
@@ -77,6 +64,8 @@ exports.findCoSupervisorIdsByThesisId = (id) => {
  * @returns ERROR: sqlite error is returned in the form {error: "message"}
  */
 exports.removeCoSupervisorsByThesisId = (thesisId) => {
+    if(!thesisId || thesisId<0)
+        throw {error: "thesisId must be greather than 0"}
     const sql = 'DELETE FROM CoSupervisorThesis WHERE id_thesis = ?';
     return new Promise((resolve, reject) => {
         db.run(sql, [thesisId], function (err) {
@@ -87,3 +76,5 @@ exports.removeCoSupervisorsByThesisId = (thesisId) => {
         });
     });
 };
+
+//==================================Support==================================
